# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-commerce API",
        "description": "API REST para sistema de e-commerce con autenticación y gestión de recursos.",
        "contact": {
            "name": "E-commerce API Support",
            "email": "api@ecommerce.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Servidor de desarrollo"
        }
    ],
    "paths": {
...
        "/api/v1/example": {
            "get": {
                "tags": [
                    "Example"
                ],
                "summary": "Obtener lista de ejemplos",
                "description": "Endpoint de ejemplo que demuestra una respuesta exitosa con datos paginados",
                "operationId": "getExamples",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Elementos por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1,
                            "example": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de ejemplos obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Ejemplo 1"
                                                            },
                                                            "description": {
                                                                "type": "string",
                                                                "example": "Descripción del ejemplo"
                                                            },
                                                            "created_at": {
                                                                "type": "string",
                                                                "format": "date-time"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/example/{id}": {
            "get": {
                "tags": [
                    "Example"
                ],
                "summary": "Obtener un ejemplo específico",
                "description": "Endpoint de ejemplo que demuestra obtener un recurso específico",
                "operationId": "getExample",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del ejemplo",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ejemplo obtenido exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Ejemplo 1"
                                                        },
                                                        "description": {
                                                            "type": "string",
                                                            "example": "Descripción del ejemplo"
                                                        },
                                                        "created_at": {
                                                            "type": "string",
                                                            "format": "date-time"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ejemplo no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/example/protected": {
            "post": {
                "tags": [
                    "Example"
                ],
                "summary": "Crear ejemplo protegido",
                "description": "Endpoint de ejemplo que requiere autenticación",
                "operationId": "createProtectedExample",
                "requestBody": {
                    "description": "Datos del ejemplo",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Nuevo ejemplo"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descripción del nuevo ejemplo"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ejemplo creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 4
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Nuevo ejemplo"
                                                        },
                                                        "description": {
                                                            "type": "string",
                                                            "example": "Descripción del nuevo ejemplo"
                                                        },
                                                        "created_at": {
                                                            "type": "string",
                                                            "format": "date-time"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/example/admin": {
            "post": {
                "tags": [
                    "Example"
                ],
                "summary": "Crear ejemplo de administrador",
                "description": "Endpoint de ejemplo que requiere rol de administrador",
                "operationId": "createAdminExample",
                "requestBody": {
                    "description": "Datos del ejemplo de admin",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Ejemplo admin"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Solo admins pueden crear esto"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ejemplo de admin creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Sin permisos de administrador",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "ApiResponse": {
                "description": "Estructura estándar de respuesta de la API",
                "properties": {
                    "success": {
                        "description": "Indica si la operación fue exitosa",
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "description": "Mensaje descriptivo de la operación",
                        "type": "string",
                        "example": "Operation completed successfully"
                    },
                    "data": {
                        "description": "Datos de respuesta (variable según endpoint)"
                    },
                    "meta": {
                        "description": "Metadatos adicionales (paginación, totales, etc.)",
                        "properties": {
                            "pagination": {
                                "properties": {
                                    "current_page": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "per_page": {
                                        "type": "integer",
                                        "example": 15
                                    },
                                    "total": {
                                        "type": "integer",
                                        "example": 150
                                    },
                                    "last_page": {
                                        "type": "integer",
                                        "example": 10
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "description": "Estructura de respuesta para errores",
                "properties": {
                    "success": {
                        "description": "Siempre false para errores",
                        "type": "boolean",
                        "example": false
                    },
                    "error": {
                        "properties": {
                            "code": {
                                "description": "Código del error",
                                "type": "string",
                                "example": "VALIDATION_ERROR"
                            },
                            "message": {
                                "description": "Mensaje del error",
                                "type": "string",
                                "example": "The given data was invalid"
                            },
                            "details": {
                                "description": "Detalles específicos del error (ej: errores de validación)",
                                "type": "object",
                                "example": {
                                    "email": [
                                        "The email field is required."
                                    ]
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "ValidationErrorResponse": {
                "description": "Respuesta específica para errores de validación (422)",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ErrorResponse"
                    }
                ]
            },
            "UnauthorizedResponse": {
                "description": "Respuesta para errores de autenticación (401)",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ErrorResponse"
                    }
                ]
            },
            "ForbiddenResponse": {
                "description": "Respuesta para errores de autorización (403)",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ErrorResponse"
                    }
                ]
            },
            "NotFoundResponse": {
                "description": "Respuesta para recursos no encontrados (404)",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ErrorResponse"
                    }
                ]
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}